+++ packages/discussions/src/utils/channels/can-edit-channel.ts
@@ -1,32 +1,26 @@
 import { IUser } from "@esri/arcgis-rest-types";
-import {
-  IChannel,
-  IDiscussionsUser,
-} from "../../types";
+import { IChannel, IDiscussionsUser, SharingAccess } from "../../types";
 import { ChannelPermission } from "../channel-permission";
-import { isOrgAdminInOrg } from "../platform";
 import { isAuthorizedToModifyChannelByLegacyPermissions } from "./is-authorized-to-modify-channel-by-legacy-permissions";
+import { hasOrgAdminUpdateRights } from "../portal-privilege";
 
 /**
- * Utility to determine if User has privileges to modify a channel
- * @deprecated use `canEditChannel` or `canDeleteChannel` instead
+ * Utility to determine if User has privileges to edit a channel
  * @param channel
  * @param user
  * @returns {boolean}
  */
-export function canModifyChannel(
+export function canEditChannel(
   channel: IChannel,
   user: IUser | IDiscussionsUser = {}
-  // channelUpdates: IUpdateChannel,
 ): boolean {
-  if (isOrgAdminInOrg(user, channel.orgId)) {
+  if (hasOrgAdminUpdateRights(user, channel.orgId)) {
     return true;
   }
 
   if (channel.channelAcl) {
     const channelPermission = new ChannelPermission(channel);
     return channelPermission.canModerateChannel(user as IDiscussionsUser);
-    // for V2: && channelPermission.canUpdateProperties(user as IDiscussionsUser, channelUpdates)
   }
 
   return isAuthorizedToModifyChannelByLegacyPermissions(user, channel);