+++ packages/discussions/src/utils/channels/index.ts
@@ -1,4 +1,6 @@
-export { canPostToChannel } from "./can-post-to-channel";
 export { canCreateChannel } from "./can-create-channel";
+export { canDeleteChannel } from "./can-delete-channel";
+export { canEditChannel } from "./can-edit-channel";
 export { canModifyChannel } from "./can-modify-channel";
+export { canPostToChannel } from "./can-post-to-channel";
 export { canReadChannel, canReadFromChannel } from "./can-read-channel";